requirements.txt
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2.3
django-graphql-extensions==0.0.6
graphene==2.1.6
graphene-django==2.3.2
graphql-core==2.2
graphql-relay==0.4.5
importlib-metadata==0.18
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
promise==2.2.1
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.2
